Move toast store to new stores dir
