+++ frontend/src/fileBrowser/FileList.tsx
@@ -3,7 +3,7 @@ import {useNavigate} from "react-router";
 import type {JSX} from "react";
 import type {ApiResponse} from "#shared/src/types.js";
 import {apiFetch} from "#/src/utils.js";
-import {useShowToast} from "#/src/layout/toastStore.ts";
+import {useShowToast} from "#/src/stores/toastStore";
 import {Button} from "../components/Button";
 
 function FileList({
